requirements.txt
asgiref==3.4.1
autopep8==1.6.0
configobj==5.0.6
dj-database-url==0.5.0
Django==3.2.9
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
Markdown==3.3.6
psycopg2==2.9.2
pycodestyle==2.8.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
validate-docbr==1.10.0
whitenoise==5.3.0
